Update packages, exception handling and adjust dependencies
The commit updates the used Java packages from javax.persistence to jakarta.persistence in several entity classes. It also updates the exception handling in the EmployeeService class to throw an EmployeeNotFoundException rather than just returning null. Furthermore, it adjusts dependencies in the pom.xml file, namely upgrading spring-boot-starter-parent version and changing the mysql-connector. It also extends the functionality of the EmployeeRepository.